Make image border optional
I noticed Matt was cropping off the borders of images (that I had spent so much time figuring out how to add).  I thought, why bother even adding them on in the first place? So I made it an optional setting that takes one additional line of code to disable.